program that finds and prints the first 98 Fibonacci numbers, starting with 1 and 2, followed by a new line. Updated again

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,7 +8,7 @@
 int main(void)
 {
 	int i;
-	float n1 = 1, n2 = 2, n3;
+	double n1 = 1, n2 = 2, n3;
 
 	printf("1, 2, ");
 	for (i = 2; i < 98; i++)
